Fix pr #5553: fix: properly serialize tool calls into dicts for token counting
